requirements.txt
azure-ai-projects
azure-identity
python-dotenv
azure-ai-agents==1.1.0b4